+++ TheGame/Models/ppo_no_lstm.py
@@ -25,6 +25,8 @@ class PPO(nn.Module):
         self.fc_v = nn.Linear(256, 1)
         self.optimizer = optim.Adam(self.parameters(), lr=learning_rate)
 
+        self.method = "PPO"
+
     def pi(self, x, softmax_dim=0):
         x = F.relu(self.fc1(x))
         x = self.fc_pi(x)