+++ TheGame/__init__.py
@@ -1,4 +1,3 @@
 from TheGame import Models
-from TheGame.Environments.Environment import GridWorld as Environment
+from TheGame.World.Environment import Environment
 from TheGame import utils
-from TheGame import Entities