+++ evolve_per.py
@@ -1,29 +1,22 @@
 from TheGame.Models.perdqn import DQNAgent
 from TheGame import Environment
+from TheGame.utils import Results
 
 # Hyperparameters
-learning_rate = 0.0005
-nr_agents = 1
-best_score = -10_000
-print_interval = 10
-scores = [0.0 for _ in range(nr_agents)]
+save_best = True
+max_epi = 20_000
+track_results = Results(print_interval=100, interactive=False)
 
 # Init env
-brains = [DQNAgent(102, 8) for _ in range(nr_agents)]
-env = Environment(width=20, height=20, nr_agents=nr_agents, evolution=True, fps=20, brains=brains, grid_size=24)
-env.max_step = 30_000
+env = Environment(width=20, height=20, nr_agents=1, evolution=True, fps=20, brains=[DQNAgent(152, 8)], grid_size=24)
 s = env.reset()
 
-for n_epi in range(2_000):
+for n_epi in range(max_epi):
     actions = [agent.brain.get_action(s[i]) for i, agent in enumerate(env.agents)]
     s_prime, r, dones, infos = env.step(actions)
 
-    # Learn only if still alive (not done)
+    # Memorize & Learn
     for i, agent in enumerate(env.agents):
-        if agent.fitness > best_score:
-            best_score = agent.fitness
-
-        # Memorize & Learn
         agent.brain.append_sample(s[i], actions[i], r[i], s_prime[i], dones[i])
         if agent.brain.memory.tree.n_entries >= agent.brain.train_start:
             agent.brain.train_model()
@@ -33,14 +26,14 @@ for n_epi in range(2_000):
         if agent.age % 20 == 0 or agent.dead:
             agent.brain.update_target_model()
 
+    track_results.update_results(env.agents, n_epi)
     s = env.update_env()
-
-    if n_epi % 100 == 0:
-        print(f"Best score: {best_score}. Nr episodes: {n_epi}. Nr_agents: {len(env.agents)}")
 
     # env.render()
 
-# s = env.reset()
+if save_best:
+    env.save_best_brain(max_epi)
+
 while True:
     actions = []
     for i, agent in enumerate(env.agents):